+++ viktor/bin/d/data/laravel-livereaction/resources/views/script.blade.php
