+++ viktor/bin/d/data/move-to-vendor.sh
