+++ viktor/bin/d/install-laravel-broadcasting.sh
