+++ viktor/bin/d/install-laravel-mercure-example.sh
