+++ viktor/bin/d/install-laravel.sh
