+++ viktor/bin/readenv.sh
