+++ viktor/cli/my-env1.env
@@ -1,0 +1,10 @@
+# viktor setup=
+DC_VIKTOR_ROOT=./viktor/
+DC_DOCKERFILE=viktor/docker-compose.yml
+DC_CONTAINER=server
+DC_SERVER_NAME=localhost
+# viktor setup
+DC_VIKTOR_ROOT=./viktor/
+DC_DOCKERFILE=viktor/docker-compose.yml
+DC_CONTAINER=server
+DC_SERVER_NAME=localhost
