+++ viktor/cli/src/docker/interact.js
